Update Lesson_4_Store_Data_On_Smart_Contract.md
In the last bit of code:
"const waveTxn = await waveContract.wave();" 

became:
"let waveTxn = await waveContract.wave();" but isn't explicitly called out. 

Not sure what the intention here is, but should that be updated to let in the other 2 places or to const in this last place? This PR does the latter.